refactor : fix class name about generator, refactor : some variable name, refactor : public -> private moveCars method in RacingGame.class
