+++ src/main/java/racing/utils/RandomGenerator.java
@@ -2,7 +2,7 @@ package racing.utils;
 
 import java.util.Random;
 
-public class Generator {
+public class RandomGenerator {
     private static final int BOUNDARY_NUMBER = 10;
 
     public static int generateNumber() {
