+++ src/test/java/racingGame/CarTest.java
@@ -2,7 +2,7 @@ package racingGame;
 
 import org.junit.Before;
 import org.junit.Test;
-import racing.Car;
+import racing.model.Car;
 
 import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.assertThat;
